MXS-1300: Add basic maxctrl tests
Added basic tests for maxctrl. They simply test that the diagnostic
commands work.

+++ test/stop_maxscale.sh
@@ -21,7 +21,7 @@ pkill -9 '^maxscale$'
 rm -r $maxscaledir/lib/maxscale
 rm -r $maxscaledir/cache/maxscale
 rm -r $maxscaledir/run/maxscale
-rm /tmp/maxadmin.sock
+test -f /tmp/maxadmin.sock && rm /tmp/maxadmin.sock
 
 mkdir -m 0755 -p $maxscaledir/lib/maxscale
 mkdir -m 0755 -p $maxscaledir/cache/maxscale
